new float_to_list/2
Attached please find a patch that adds a new float_to_list/2 BIF.  The
patch was created off of the master branch of https://github.com/erlang/otp.

This BIF solves a problem of float_to_list/1 that doesn't allow
specifying the number of digits after the decimal point when formatting
floats.

	float_to_list(Float, Options) -> string()

	Float = float()
	Options = [Option]
	Option = {precision, Precision::integer()} | compact

	Text representation of a float formatted using given options

	Returns a string which corresponds to the text
	representation of Float using fixed decimal point formatting.
         When precision option is specified
         the returned value will contain at most Precision number of
         digits past the decimal point.  When compact option is provided
         the trailing zeros at the end of the list are truncated.

	1> float_to_list(7.12, [{precision, 4}]).
	"7.1200"
	2> float_to_list(7.12, [{precision, 4}, compact]).
	"7.12"

Documentation and test cases are updated to reflect the changes.