Add gen_stream behaviour
This is a new OTP behaviour designed to serve serial streams
as an ordered sequence of "blocks" of data.  The stream is
loaded concurrently to a process-striped internal buffer so
that slow streams such as disk files can be efficiently
delivered. The internal buffering can be massaged by
supplying a transform function to manipulate the data blocks
between the source and the internal buffers to perform
such tasks as chopping lines of text, compressing /
decompressing or encrypting / unencrypting.

The abstraction also allows the creation of functionally
generated streams and infinite streams.  Both of these
cases allow for continuous streaming without a large
memory footprint.

This patch includes a new module in stdlib (gen_stream.erl),
a new test suite (gen_stream_SUITE.erl) and a new page in
the documentation (gen_stream.xml).